Starting to test functions of GP2xCommands - test setOffCardKeys(), setOffCardKey(), deleteOffCardKey(), getKey(), getKeys() - switch to byte type the parameters of deleteOffCardKey()

+++ src/main/java/fr/xlim/ssd/opal/library/GPApplet.java
@@ -111,7 +111,7 @@ public class GPApplet {
      * @param keyId
      * @return
      */
-    public SCKey deleteOffCardKey(int keySetVersion, int keyId) {
+    public SCKey deleteOffCardKey(byte keySetVersion, byte keyId) {
         return this.cmds.deleteOffCardKey(keySetVersion, keyId);
     }
